UHM-4445: Migrate Haiti, Liberia, and Sierra Leone address bundles to their respective config files
