--- api/src/main/resources/addresshierarchy/sierra_leone_address_hierarchy_entries_4.csv
@@ -1,210 +1,0 @@
-Guinea|-|-|-
-Liberia|-|-|-
-Sierra Leone|Bombali|Bombali Shebora|-|
-Sierra Leone|Bombali|Makari Gbanti|-|
-Sierra Leone|Bombali|Sanda Tenraren|-|
-Sierra Leone|Bombali|Biriwa|-|
-Sierra Leone|Bombali|Magbaiamba Ndowahun|-|
-Sierra Leone|Bombali|Sanda Loko|-|
-Sierra Leone|Bombali|Safroko Limba|-|
-Sierra Leone|Bombali|Sella Limba|-|
-Sierra Leone|Bombali|Tambakha|-|
-Sierra Leone|Bombali|Gbanti Kamaranka|-|
-Sierra Leone|Bombali|Gbendembu Ngowahun|-|
-Sierra Leone|Bombali|Paki Massabong|-|
-Sierra Leone|Bombali|Libeisaygahun|-|
-Sierra Leone|Kambia|Magbema|-|
-Sierra Leone|Kambia|Samu|-|
-Sierra Leone|Kambia|Mambolo|-|
-Sierra Leone|Kambia|Gbinle Dixing|-|
-Sierra Leone|Kambia|Masungbala|-|
-Sierra Leone|Kambia|Tonko Limba|-|
-Sierra Leone|Kambia|Briama|-|
-Sierra Leone|Koinadugu|Diang|-|
-Sierra Leone|Koinadugu|Nieni|-|
-Sierra Leone|Koinadugu|Sengbe|-|
-Sierra Leone|Koinadugu|Sulima|-|
-Sierra Leone|Koinadugu|Wara-Wara Bafodea|-|
-Sierra Leone|Koinadugu|Mongo|-|
-Sierra Leone|Koinadugu|Dembelia Sikunia|-|
-Sierra Leone|Koinadugu|Kasunko|-|
-Sierra Leone|Koinadugu|Folasaba|-|
-Sierra Leone|Koinadugu|Wara-Wara Yagala|-|
-Sierra Leone|Koinadugu|Neya|-|
-Sierra Leone|Port Loko|Koya|-|
-Sierra Leone|Port Loko|Kaffu Bullom|-|
-Sierra Leone|Port Loko|T.M. Safroko|-|
-Sierra Leone|Port Loko|Loko Massama|-|
-Sierra Leone|Port Loko|Maforki|-|
-Sierra Leone|Port Loko|Masimera|-|
-Sierra Leone|Port Loko|Bureh Kaseh Maconteh|-|
-Sierra Leone|Port Loko|Marampa|-|
-Sierra Leone|Port Loko|Buya Romende|-|
-Sierra Leone|Port Loko|Debia|-|
-Sierra Leone|Port Loko|Sanda Magbolontor|-|
-Sierra Leone|Tonkolili|Kalansongoia|-|
-Sierra Leone|Tonkolili|Gbonkolenken|-|
-Sierra Leone|Tonkolili|Kholifa|-|
-Sierra Leone|Tonkolili|Kunike|-|
-Sierra Leone|Tonkolili|Kafe Simiria|-|
-Sierra Leone|Tonkolili|Kholifa Mabang|-|
-Sierra Leone|Tonkolili|Kunike Barina|-|
-Sierra Leone|Tonkolili|Malal Mara|-|
-Sierra Leone|Tonkolili|Sambaia|-|
-Sierra Leone|Tonkolili|Yoni|-|
-Sierra Leone|Tonkolili|Tane|-|
-Sierra Leone|Bo|Badjia|-|
-Sierra Leone|Bo|Wonde|-|
-Sierra Leone|Bo|Selenga|-|
-Sierra Leone|Bo|Niawa Lenga|-|
-Sierra Leone|Bo|Komboya|-|
-Sierra Leone|Bo|Bagbwe|-|
-Sierra Leone|Bo|Valunia|-|
-Sierra Leone|Bo|Jaiama Bongor|-|
-Sierra Leone|Bo|Bagbo|-|
-Sierra Leone|Bo|Tikonko|-|
-Sierra Leone|Bo|Kakua|-|
-Sierra Leone|Bo|Bumpe-Gao|-|
-Sierra Leone|Bo|Baoma|-|
-Sierra Leone|Bo|Gbo|-|
-Sierra Leone|Bo|Lugbu|-|
-Sierra Leone|Bonthe|Imperri|-|
-Sierra Leone|Bonthe|Bum|-|
-Sierra Leone|Bonthe|Jong|-|
-Sierra Leone|Bonthe|Dema|-|
-Sierra Leone|Bonthe|Kwamebai Krim|-|
-Sierra Leone|Bonthe|Sittia|-|
-Sierra Leone|Bonthe|Nongoba Bullom|-|
-Sierra Leone|Bonthe|Sogbini|-|
-Sierra Leone|Bonthe|Yawbeko|-|
-Sierra Leone|Bonthe|Kpanda Kemo|-|
-Sierra Leone|Bonthe|Bendu Cha|-|
-Sierra Leone|Moyamba|Bumpe|-|
-Sierra Leone|Moyamba|Kowa|-|
-Sierra Leone|Moyamba|Kamajei|-|
-Sierra Leone|Moyamba|Fakunya|-|
-Sierra Leone|Moyamba|Dasse|-|
-Sierra Leone|Moyamba|Upper Banta|-|
-Sierra Leone|Moyamba|Kagboro|-|
-Sierra Leone|Moyamba|Banta|-|
-Sierra Leone|Moyamba|Bagruna|-|
-Sierra Leone|Moyamba|Kaiyamba|-|
-Sierra Leone|Moyamba|Timdale|-|
-Sierra Leone|Moyamba|Ribbi|-|
-Sierra Leone|Moyamba|Kongbora|-|
-Sierra Leone|Moyamba|Fori|-|
-Sierra Leone|Pujehun|Peje|-|
-Sierra Leone|Pujehun|Panga Krim|-|
-Sierra Leone|Pujehun|Mano Sakrim|-|
-Sierra Leone|Pujehun|Malen|-|
-Sierra Leone|Pujehun|Makpele|-|
-Sierra Leone|Pujehun|Kpaka|-|
-Sierra Leone|Pujehun|Soro Gbema|-|
-Sierra Leone|Pujehun|Yekomo Kpukumu Krim|-|
-Sierra Leone|Pujehun|Kpanga Kagonde|-|
-Sierra Leone|Pujehun|Gallines Perri|-|
-Sierra Leone|Pujehun|Barri|-|
-Sierra Leone|Pujehun|Sowa|-|
-Sierra Leone|Kailahun|Peje West|-|
-Sierra Leone|Kailahun|Jaluahun|-|
-Sierra Leone|Kailahun|Jawei|-|
-Sierra Leone|Kailahun|Luawa|-|
-Sierra Leone|Kailahun|Malema|-|
-Sierra Leone|Kailahun|Kissi Teng|-|
-Sierra Leone|Kailahun|Kissi Tongi|-|
-Sierra Leone|Kailahun|Kissi Kama|-|
-Sierra Leone|Kailahun|Peje Bongre|-|
-Sierra Leone|Kailahun|Upper Bambara|-|
-Sierra Leone|Kailahun|Yawei|-|
-Sierra Leone|Kailahun|Dea|-|
-Sierra Leone|Kailahun|Penguia|-|
-Sierra Leone|Kailahun|Mandu|-|
-Sierra Leone|Kenema|Dama|-|
-Sierra Leone|Kenema|Tunkia|-|
-Sierra Leone|Kenema|Nongowa|-|
-Sierra Leone|Kenema|Lower Bambara|-|
-Sierra Leone|Kenema|Kandu Leppiam|-|
-Sierra Leone|Kenema|Gaura|-|
-Sierra Leone|Kenema|Gorama Mende|-|
-Sierra Leone|Kenema|Niawa|-|
-Sierra Leone|Kenema|Koya|-|
-Sierra Leone|Kenema|Malegohun|-|
-Sierra Leone|Kenema|Simbaru|-|
-Sierra Leone|Kenema|Small Bo|-|
-Sierra Leone|Kenema|Dodo|-|
-Sierra Leone|Kenema|Wando|-|
-Sierra Leone|Kenema|Nomo|-|
-Sierra Leone|Kenema|Langurama|-|
-Sierra Leone|Kono|Tankoro|Tankoro Section
-Sierra Leone|Kono|Tankoro|Woafeh Section
-Sierra Leone|Kono|Tankoro|Njama Section
-Sierra Leone|Kono|Gbense|Moindekor Section
-Sierra Leone|Kono|Gbense|Moindefeh Section
-Sierra Leone|Kono|Gbense|Bayafeh Section
-Sierra Leone|Kono|Gbense|Bayakor Section
-Sierra Leone|Kono|Gbense|Bafinfeh Section
-Sierra Leone|Kono|Kamara|Kongorfinjah Section
-Sierra Leone|Kono|Kamara|Gbondu Section
-Sierra Leone|Kono|Kamara|Sukudu Section
-Sierra Leone|Kono|Kamara|Dangbaidu Section
-Sierra Leone|Kono|Nimikoro|Jaiama Section
-Sierra Leone|Kono|Nimikoro|Bafinfeh Section
-Sierra Leone|Kono|Nimikoro|Bandafafeh Section
-Sierra Leone|Kono|Nimikoro|Gbogboafeh Section
-Sierra Leone|Kono|Nimikoro|Mansayifeh Section
-Sierra Leone|Kono|Nimiyama|Njaiafeh Section
-Sierra Leone|Kono|Nimiyama|Bafinfeh Section
-Sierra Leone|Kono|Nimiyama|Njagbakahun Section
-Sierra Leone|Kono|Nimiyama|Peyifeh Section
-Sierra Leone|Kono|Nimiyama|Taama Forest Section
-Sierra Leone|Kono|Sandor|Kawafeh Section
-Sierra Leone|Kono|Sandor|Sumunjifeh Section
-Sierra Leone|Kono|Sandor|Sinkongoefeh Section
-Sierra Leone|Kono|Sandor|Njeikor Section
-Sierra Leone|Kono|Sandor|Sangbafeh Section
-Sierra Leone|Kono|Sandor|Fakongor Section
-Sierra Leone|Kono|Sandor|Bafinfeh Section
-Sierra Leone|Kono|Sandor|Yawantanda Section
-Sierra Leone|Kono|Sandor|Dangbaidu Section
-Sierra Leone|Kono|Toli|Kwidu Section
-Sierra Leone|Kono|Toli|Komadu Section
-Sierra Leone|Kono|Toli|Bawadu Section
-Sierra Leone|Kono|Gbane|Mongor Section
-Sierra Leone|Kono|Gbane|Kamaah Section
-Sierra Leone|Kono|Gbane|Gbikidakor Section
-Sierra Leone|Kono|Gbane|Mahenkandor Section
-Sierra Leone|Kono|Gbane|Gbane Yemaor Section
-Sierra Leone|Kono|Lei|Lei Section
-Sierra Leone|Kono|Lei|Keinsay Section
-Sierra Leone|Kono|Lei|Kamara Section
-Sierra Leone|Kono|Lei|Sandor Section
-Sierra Leone|Kono|Lei|Tankoro Section
-Sierra Leone|Kono|Lei|Sangbadaa Section
-Sierra Leone|Kono|Lei|Koaro Section
-Sierra Leone|Kono|Lei|Dea Section
-Sierra Leone|Kono|Lei|Yawai Section
-Sierra Leone|Kono|Soa|Koo-Kongorkuma Section
-Sierra Leone|Kono|Soa|Morfinkor Section
-Sierra Leone|Kono|Soa|Tensedakor Section
-Sierra Leone|Kono|Soa|Mofinkor Section
-Sierra Leone|Kono|Soa|Tensekor Section
-Sierra Leone|Kono|Soa|Mongor Section
-Sierra Leone|Kono|Soa|Sawa Fiama Section
-Sierra Leone|Kono|Soa|Sawa Buuma Section
-Sierra Leone|Kono|Soa|Maindu Section
-Sierra Leone|Kono|Fiama|Fiama Section
-Sierra Leone|Kono|Fiama|Dumbia Section
-Sierra Leone|Kono|Fiama|Yorkor Section
-Sierra Leone|Kono|Fiama|Kooma Section
-Sierra Leone|Kono|Fiama|Kookaa Section
-Sierra Leone|Kono|Gorama Kono|Kangama Section
-Sierra Leone|Kono|Gorama Kono|Bunabu Section
-Sierra Leone|Kono|Gorama Kono|Selokoma Section
-Sierra Leone|Kono|Mafindor|Kamiendor Section
-Sierra Leone|Kono|Mafindor|Kutay Section
-Sierra Leone|Kono|Mafindor|Mafindor Section
-Sierra Leone|Kono|Gbane Kandor|Gbane Tetehmah Section
-Sierra Leone|Kono|Gbane Kandor|Gbaneh Kuor Section
-Sierra Leone|Kono|Gbane Kandor|Gbendekor Section
-Sierra Leone|Kono|Gbane Kandor|Dambidu Section
-Sierra Leone|Kono|Gbane Kandor|Gbanekandor Section
