+++ api/src/test/resources/test_address_hierarchy_entries.csv
@@ -1,0 +1,6 @@
+Yon lot peyi|-|-|-|Yon lot
+Haiti|Ouest|Port-au-Prince|6ème Turgeau^111-01|Deprez
+Haiti|Ouest|Port-au-Prince|6ème Turgeau^111-01|Boutilier
+Haiti|Ouest|Cite Soleil|1ère Varreux^117-01|Sarthe
+Haiti|Ouest|Cite Soleil|1ère Varreux^117-01|Menelas
+Haiti|Ouest|Cite Soleil|2ème Varreux^117-02|Blanchard
